% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{as.character.centralperk}
\alias{as.character.centralperk}
\title{S3 as.character method for 'centralperk' class}
\usage{
\method{as.character}{centralperk}(x, ...)
}
\description{
S3 as.character method for 'centralperk' class
}
