% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{centralperk}
\alias{centralperk}
\alias{ross}
\alias{rachel}
\alias{monica}
\alias{joey}
\alias{phoebe}
\alias{chandler}
\title{Get a random 'Friends' Quote}
\usage{
centralperk()

ross()

rachel()

monica()

joey()

phoebe()

chandler()
}
\value{
list of class 'centralperk' with 2 elements: character and quote
}
\description{
Get a random 'Friends' Quote
}
\section{Functions}{
\itemize{
\item \code{centralperk}: Random quote from a random main character from Friends

\item \code{ross}: A random quote from Ross

\item \code{rachel}: A random quote from Rachel

\item \code{monica}: A random quote from Monica

\item \code{joey}: A random quote from Joey

\item \code{phoebe}: A random quote from Phoebe

\item \code{chandler}: A random quote from Chandler
}}

\examples{
centralperk()
}
