% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{friends_quotes_api}
\alias{friends_quotes_api}
\title{Main interface function to "Friends Quotes API"}
\usage{
friends_quotes_api(path)
}
\arguments{
\item{path}{Parameter for API. Leave blank, any integer, or 'random'}
}
\value{
list of class 'friends_quotes_api' with API response
}
\description{
Main interface function to "Friends Quotes API"
}
\keyword{internal}
