% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.centralperk}
\alias{print.centralperk}
\title{S3 print method for 'centralperk' class}
\usage{
\method{print}{centralperk}(x, ...)
}
\description{
S3 print method for 'centralperk' class
}
\keyword{internal}
