% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.friends_quotes_api}
\alias{print.friends_quotes_api}
\title{S3 print method for 'friends_quotes_api' class}
\usage{
\method{print}{friends_quotes_api}(x, ...)
}
\description{
S3 print method for 'friends_quotes_api' class
}
\keyword{internal}
