% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{single_character}
\alias{single_character}
\title{Grab quote for a single character}
\usage{
single_character(character)
}
\arguments{
\item{character}{Name of main Friends character}
}
\value{
list of class 'centralperk' with 2 elements: character and quote
}
\description{
Grab quote for a single character
}
\keyword{internal}
